backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0

sentence-transformers==2.2.2
faiss-cpu==1.7.4

ollama==0.1.2

pydantic==1.10.14
python-multipart==0.0.9

# Optional but recommended for performance
numpy==1.26.4

# Only if your rag.py requires jsonlines (usually not, but safe)
jsonlines==3.1.0
# Auth
google-auth==2.28.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pyjwt==2.10.1
bcrypt==4.1.2